+++ src/tightdb/objc/test/shared_group.m
@@ -50,7 +50,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     TDBContext* fromDisk = [TDBContext contextWithPersistenceToFile:@"employees.tightdb" error:nil];
 
     [fromDisk readWithBlock:^(TDBTransaction* group) {
-            SharedTable2* diskTable = [group getTableWithName:@"employees" asTableClass:[SharedTable2 class]];
+            SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < [diskTable rowCount]; i++) {
                 SharedTable2Row *cursor = [diskTable rowAtIndex:i];
@@ -62,7 +62,7 @@ TIGHTDB_TABLE_2(SharedTable2,
 
 
     [fromDisk writeWithBlock:^(TDBTransaction* group) {
-            SharedTable2* diskTable = [group getTableWithName:@"employees" asTableClass:[SharedTable2 class]];
+            SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < 50; i++) {
                 [diskTable addHired:YES Age:i];
@@ -72,7 +72,7 @@ TIGHTDB_TABLE_2(SharedTable2,
 
 
     [fromDisk writeWithBlock:^(TDBTransaction* group) {
-            SharedTable2* diskTable = [group getTableWithName:@"employees" asTableClass:[SharedTable2 class]];
+            SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < 50; i++) {
                 [diskTable addHired:YES Age:i];
@@ -82,19 +82,19 @@ TIGHTDB_TABLE_2(SharedTable2,
 
 
     [fromDisk writeWithBlock:^(TDBTransaction* group) {
-            SharedTable2* diskTable = [group getTableWithName:@"employees" asTableClass:[SharedTable2 class]];
+            SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < 50; i++) {
                 [diskTable addHired:YES Age:i];
             }
         
-            STAssertNil([group getTableWithName:@"Does not exist"], @"Table does not exist");
+            STAssertNil([group tableWithName:@"Does not exist"], @"Table does not exist");
 
             return YES; // commit
         } error:nil];
 
     [fromDisk readWithBlock:^(TDBTransaction* group) {
-            SharedTable2* diskTable = [group getTableWithName:@"employees" asTableClass:[SharedTable2 class]];
+            SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
         
         STAssertThrows([diskTable removeAllRows], @"Not allowed in readtransaction");
@@ -126,7 +126,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     } error:nil];
     
     [fromDisk readWithBlock:^(TDBTransaction* group) {
-        TDBTable *t = [group getTableWithName:@"table"];
+        TDBTable *t = [group tableWithName:@"table"];
        
         TDBQuery *q = [t where];
         
@@ -139,7 +139,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         STAssertTrue([q countRows] == 1, @"No rows have been removed");
         STAssertTrue([v rowCount] == 1, @"No rows have been removed");
         
-        STAssertNil([group getTableWithName:@"Does not exist"], @"Table does not exist");
+        STAssertNil([group tableWithName:@"Does not exist"], @"Table does not exist");
     }];
 }
 
@@ -196,7 +196,7 @@ TIGHTDB_TABLE_2(SharedTable2,
 
     
     [sg writeWithBlock:^(TDBTransaction* group) {
-        TDBTable *t = [group getTableWithName:@"t"];
+        TDBTable *t = [group tableWithName:@"t"];
         [t addColumnWithName:@"col" andType:TDBBoolType];
         NSUInteger rowIndex = [t addRow:nil];
         TDBRow *row = [t rowAtIndex:rowIndex];
@@ -212,7 +212,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     
     
     [sg2 writeWithBlock:^(TDBTransaction* group) {
-        TDBTable *t = [group getTableWithName:@"t"];
+        TDBTable *t = [group tableWithName:@"t"];
         [t addRow:nil]; /* Adding an empty row */
         return YES;
     } error:nil];
@@ -243,12 +243,115 @@ TIGHTDB_TABLE_2(SharedTable2,
     
     [c readWithBlock:^(TDBTransaction *transaction) {
         
-        STAssertThrows([transaction getTableWithName:nil], @"name is nil");
-        STAssertThrows([transaction getTableWithName:@""], @"name is empty");
+        STAssertThrows([transaction tableWithName:nil], @"name is nil");
+        STAssertThrows([transaction tableWithName:@""], @"name is empty");
         STAssertThrows([transaction createTableWithName:@"same name"], @"creating table not allowed in read transaction");
         STAssertThrows([transaction createTableWithName:@"name"], @"creating table not allowed in read transaction");
-        STAssertNil([transaction getTableWithName:@"weird name"], @"get table that does not exists return nil");
+        STAssertNil([transaction tableWithName:@"weird name"], @"get table that does not exists return nil");
     }];
+}
+
+-(void)testPinnedTransactions
+{
+    NSString *contextPath = @"pinnedTransactions.tightdb";
+    NSFileManager* fm = [NSFileManager defaultManager];
+    [fm removeItemAtPath:contextPath error:nil];
+    [fm removeItemAtPath:[contextPath stringByAppendingString:@".lock"] error:nil];
+   
+    TDBContext *context1 = [TDBContext contextWithPersistenceToFile:contextPath error:nil];
+    TDBContext *context2 = [TDBContext contextWithPersistenceToFile:contextPath error:nil];
+    
+    {
+        // initially, always say that the db has changed
+        BOOL changed = [context2 pinReadTransactions];
+        STAssertTrue(changed, nil);
+        [context2 unpinReadTransactions];
+        // asking again - this time there is no change
+        changed = [context2 pinReadTransactions];
+        STAssertFalse(changed, nil);
+
+        [context2 unpinReadTransactions];
+    }
+    {   // add something to the db to play with
+        [context1 writeWithBlock:^BOOL(TDBTransaction *transaction) {
+            TDBTable *t1 = [transaction createTableWithName:@"test"];
+            [t1 addColumnWithName:@"col0" andType:TDBBoolType];
+            [t1 addRow:@[@YES]];
+            //t1->add(0, 2, false, "test");
+            return YES;
+        } error:nil];
+    }
+    {   // validate that we can see previous commit from within a new pinned transaction
+        BOOL changed = [context2 pinReadTransactions];
+        STAssertTrue(changed, nil);
+        [context2 readWithBlock:^(TDBTransaction *transaction) {
+            TDBTable *t = [transaction tableWithName:@"test"];
+            STAssertEquals([[t rowAtIndex:0] boolInColumnWithIndex:0], YES, nil);
+        }];
+    }
+    {   // commit new data in another context, without unpinning
+        [context1 writeWithBlock:^BOOL(TDBTransaction *transaction) {
+            TDBTable *t = [transaction tableWithName:@"test"];
+            [t addRow:@[@NO]];
+            return YES;
+        } error:nil];
+        
+    }
+    {   // validate that we can see previous commit if we're not pinned
+        [context1 readWithBlock:^(TDBTransaction *transaction) {
+            TDBTable *t = [transaction tableWithName:@"test"];
+            STAssertEquals([[t rowAtIndex:1] boolInColumnWithIndex:0], NO, nil);
+        }];
+        
+    }
+     {   // validate that we can NOT see previous commit from within a pinned transaction
+        [context2 readWithBlock:^(TDBTransaction *transaction) {
+            TDBTable *t = [transaction tableWithName:@"test"];
+            STAssertEquals(t.rowCount, (NSUInteger)1, @"Still only 1 row");
+        }];
+        
+    }
+    {   // unpin, pin again and validate that we can now see previous commit
+        [context2 unpinReadTransactions];
+        BOOL changed = [context2 pinReadTransactions];
+        STAssertTrue(changed, @"changes since last transaction");
+        [context2 readWithBlock:^(TDBTransaction *transaction) {
+            TDBTable *t = [transaction tableWithName:@"test"];
+            STAssertEquals(t.rowCount, (NSUInteger)2, @"Now we see 2 rows");
+            STAssertEquals([[t rowAtIndex:1] boolInColumnWithIndex:0], NO, nil);
+        }];
+    }
+    {   // can't pin if already pinned
+        STAssertThrows([context2 pinReadTransactions], @"Already pinned");
+    }
+    {   // can't unpin if already unpinned
+        [context2 unpinReadTransactions];
+        STAssertThrows([context2 unpinReadTransactions], @"Already unpinned");
+
+    }
+    {   // can't pin while we're inside a transaction
+        [context1 readWithBlock:^(TDBTransaction *transaction) {
+            STAssertThrows([context1 pinReadTransactions], @"Can't pin inside transaction");
+            STAssertNotNil(transaction, @"Parameter must be used");
+        }];
+    }
+    
+    {   // can't unpin while we're inside a transaction
+        [context1 pinReadTransactions];
+        [context1 readWithBlock:^(TDBTransaction *transaction) {
+            STAssertThrows([context1 unpinReadTransactions], @"Can't unpin inside transaction");
+            STAssertNotNil(transaction, @"Parameter must be used");
+        }];
+        [context1 unpinReadTransactions];
+    }
+    {   // can't start a write transaction while pinned
+        [context1 pinReadTransactions];
+        STAssertThrows([context1 writeWithBlock:^BOOL(TDBTransaction *transaction) {
+            STAssertNotNil(transaction, @"Parameter must be used");
+            return YES;
+        } error:nil], @"Can't start write transaction while pinned");
+        [context1 unpinReadTransactions];
+    }
 }
 
 @end